Moving the self-calibration to a subfolder
