+++ txpipe/__init__.py
@@ -32,4 +32,6 @@ from .metadata import TXTracerMetadata
 from .convergence import TXConvergenceMaps
 from .map_correlations import TXMapCorrelations
 from .rail_pz import PZRailTrain
-from .twopoint_scia import TXSelfCalibrationIA
+# Here are the stages that mostly will be used for other projects
+# such as the self-calibration of Intrinsic alignment.
+from .extensions.twopoint_scia import TXSelfCalibrationIA
